requirements/common.txt
Django==4.0.5
psycopg2-binary==2.9.3
django-filter==21.1
djangorestframework==3.13.1
markdown==3.3.7